Перемещение модалок для выбора сервера в отдельную папку
